upgrading project to php 8.1

+++ src/Util/AnimateXY.php
@@ -7,16 +7,19 @@ namespace Trismegiste\Genetic\Util;
  *
  * @author flo
  */
-class AnimateXY extends ImagickPlotter {
+class AnimateXY extends ImagickPlotter
+{
 
     protected $pattern;
 
-    public function __construct(int $width, int $height, string $patternName) {
+    public function __construct(int $width, int $height, string $patternName)
+    {
         parent::__construct($width, $height);
         $this->pattern = $patternName;
     }
 
-    public function draw(array& $data) {
+    public function draw(array & $data)
+    {
         $extrem = $this->getBoundaries($data);
 
         $deltaX = $extrem->infX;
@@ -30,18 +33,18 @@ class AnimateXY extends ImagickPlotter {
             $handle = $this->createImage($this->width, $this->height, 0, 0, 0);
             $hue = 270.0 * $idx / $curvesCount;
             $rgb = $this->hsv2rgb($hue, 1.0, 1.0);
-            $plotColor = imagecolorallocate($handle, $rgb[0], $rgb[1], $rgb[2]);
+            $plotColor = imagecolorallocate($handle, (int) $rgb[0], (int) $rgb[1], (int) $rgb[2]);
             // legend
             $step = floor($idx / $curvesCount * 10) / 10;
             if ($lastStep !== $step) {
-                imagefttext($handle, $this->height / 50, 0, $this->width * 0.92, $this->height * ($step + 1 / 20), $plotColor, './bin/akukamu.otf', "Curve $idx");
+                imagefttext($handle, (int) ($this->height / 50), 0, (int) ($this->width * 0.92), (int) ($this->height * ($step + 1 / 20)), $plotColor, './bin/akukamu.otf', "Curve $idx");
                 $lastStep = $step;
             }
             // plotting
             foreach ($generation as $plot) {
                 $x = $this->width * 0.05 + ($plot['x'] - $deltaX) / $scaleX;
                 $y = $this->height - ($this->height * 0.05 + ($plot['y'] - $deltaY) / $scaleY);
-                imagefilledellipse($handle, $x, $y, 4, 4, $plotColor);
+                imagefilledellipse($handle, (int) $x, (int) $y, 4, 4, $plotColor);
             }
             imagecolordeallocate($handle, $plotColor);
             imagepng($handle, sprintf($this->pattern, $idx));
